+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - AssetsService.GetAssetParams
 - new proto файлы (Версия 2.4.0 (05.06.2025))
 - client.NewAccountRequest(accountId string).Do(ctx) = обертка для более "простого" получения данных по счету  
+- client.NewAccountTradesRequest(accountId string).Do(ctx) = обертка для более "простого" получения сделок по счету
 - client.GetTime(ctx context.Context) (time.Time, error)
 - 
 ## [0.1.0] - 2025-04-17